+++ models/user_channel.go
@@ -5,8 +5,8 @@ import (
 )
 
 type UserChannel struct {
-	UserID    uint      `gorm:"primaryKey" json:"user_id"`
-	ChannelID uint      `gorm:"primaryKey" json:"channel_id"`
+	UserID    uint64    `gorm:"primaryKey" json:"user_id"`
+	ChannelID uint64    `gorm:"primaryKey" json:"channel_id"`
 	JoinedAt  time.Time `gorm:"autoCreateTime;column:joined_at" json:"joined_at"`
 
 	User    User    `gorm:"foreignKey:UserID;constraint:OnDelete:CASCADE" json:"-"`
